// frontend-angular/src/proto/driver.ts
export interface RegisterDriverRequest {
  userId: string;
  vehicleNo: string;
  capacity: number;
  model: string;
  color: string;
}

export interface DriverProfile {
  driverId: string;
  userId: string;
  vehicleNo: string;
  capacity: number;
  routes: RoutePlan[];
  model: string;
  color: string;
}

export interface RoutePlan {
  routeId: string;
  driverId: string;
  stops: RouteStop[];
  finalAreaId: string;
  createdAt: Date | undefined;
}

export interface RouteStop {
  sequence: number;
  areaId: string;
  isStation: boolean;
  arrivalOffsetMinutes: number;
}

export interface RegisterRouteRequest {
  driverId: string;
  stops: RouteStop[];
}

export interface UpdateRouteRequest {
  driverId: string;
  routeId: string;
  stops: RouteStop[];
}

export interface UpdatePickupRequest {
  driverId: string;
  routeId: string;
  pickingUp: boolean;
}

export interface Ack {
  ok: boolean;
  msg: string;
}

export interface DriverId {
  id: string;
}

function createBaseRegisterDriverRequest(): RegisterDriverRequest {
  return { userId: "", vehicleNo: "", capacity: 0, model: "", color: "" };
}

export const RegisterDriverRequest: MessageFns<RegisterDriverRequest> = {
  encode(message: RegisterDriverRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(10).string(message.userId);
    }
    if (message.vehicleNo !== "") {
      writer.uint32(18).string(message.vehicleNo);
    }
    if (message.capacity !== 0) {
      writer.uint32(24).int32(message.capacity);
    }
    if (message.model !== "") {
      writer.uint32(34).string(message.model);
    }
    if (message.color !== "") {
      writer.uint32(42).string(message.color);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RegisterDriverRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterDriverRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.vehicleNo = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.capacity = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.model = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.color = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RegisterDriverRequest {
    return {
      userId: isSet(object.userId) ? globalThis.String(object.userId) : "",
      vehicleNo: isSet(object.vehicleNo) ? globalThis.String(object.vehicleNo) : "",
      capacity: isSet(object.capacity) ? globalThis.Number(object.capacity) : 0,
      model: isSet(object.model) ? globalThis.String(object.model) : "",
      color: isSet(object.color) ? globalThis.String(object.color) : "",
    };
  },

  toJSON(message: RegisterDriverRequest): unknown {
    const obj: any = {};
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.vehicleNo !== "") {
      obj.vehicleNo = message.vehicleNo;
    }
    if (message.capacity !== 0) {
      obj.capacity = Math.round(message.capacity);
    }
    if (message.model !== "") {
      obj.model = message.model;
    }
    if (message.color !== "") {
      obj.color = message.color;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RegisterDriverRequest>, I>>(base?: I): RegisterDriverRequest {
    return RegisterDriverRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RegisterDriverRequest>, I>>(object: I): RegisterDriverRequest {
    const message = createBaseRegisterDriverRequest();
    message.userId = object.userId ?? "";
    message.vehicleNo = object.vehicleNo ?? "";
    message.capacity = object.capacity ?? 0;
    message.model = object.model ?? "";
    message.color = object.color ?? "";
    return message;
  },
};

function createBaseDriverProfile(): DriverProfile {
  return { driverId: "", userId: "", vehicleNo: "", capacity: 0, routes: [], model: "", color: "" };
}

export const DriverProfile: MessageFns<DriverProfile> = {
  encode(message: DriverProfile, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.driverId !== "") {
      writer.uint32(10).string(message.driverId);
    }
    if (message.userId !== "") {
      writer.uint32(18).string(message.userId);
    }
    if (message.vehicleNo !== "") {
      writer.uint32(26).string(message.vehicleNo);
    }
    if (message.capacity !== 0) {
      writer.uint32(32).int32(message.capacity);
    }
    for (const v of message.routes) {
      RoutePlan.encode(v!, writer.uint32(42).fork()).join();
    }
    if (message.model !== "") {
      writer.uint32(50).string(message.model);
    }
    if (message.color !== "") {
      writer.uint32(58).string(message.color);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DriverProfile {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDriverProfile();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.driverId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.vehicleNo = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.capacity = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.routes.push(RoutePlan.decode(reader, reader.uint32()));
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.model = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.color = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DriverProfile {
    return {
      driverId: isSet(object.driverId) ? globalThis.String(object.driverId) : "",
      userId: isSet(object.userId) ? globalThis.String(object.userId) : "",
      vehicleNo: isSet(object.vehicleNo) ? globalThis.String(object.vehicleNo) : "",
      capacity: isSet(object.capacity) ? globalThis.Number(object.capacity) : 0,
      routes: globalThis.Array.isArray(object?.routes) ? object.routes.map((e: any) => RoutePlan.fromJSON(e)) : [],
      model: isSet(object.model) ? globalThis.String(object.model) : "",
      color: isSet(object.color) ? globalThis.String(object.color) : "",
    };
  },

  toJSON(message: DriverProfile): unknown {
    const obj: any = {};
    if (message.driverId !== "") {
      obj.driverId = message.driverId;
    }
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.vehicleNo !== "") {
      obj.vehicleNo = message.vehicleNo;
    }
    if (message.capacity !== 0) {
      obj.capacity = Math.round(message.capacity);
    }
    if (message.routes?.length) {
      obj.routes = message.routes.map((e) => RoutePlan.toJSON(e));
    }
    if (message.model !== "") {
      obj.model = message.model;
    }
    if (message.color !== "") {
      obj.color = message.color;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DriverProfile>, I>>(base?: I): DriverProfile {
    return DriverProfile.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DriverProfile>, I>>(object: I): DriverProfile {
    const message = createBaseDriverProfile();
    message.driverId = object.driverId ?? "";
    message.userId = object.userId ?? "";
    message.vehicleNo = object.vehicleNo ?? "";
    message.capacity = object.capacity ?? 0;
    message.routes = object.routes?.map((e) => RoutePlan.fromPartial(e)) || [];
    message.model = object.model ?? "";
    message.color = object.color ?? "";
    return message;
  },
};
// ...
export interface DriverService {
  RegisterDriver(request: RegisterDriverRequest): Promise<DriverProfile>;
  RegisterRoute(request: RegisterRouteRequest): Promise<RoutePlan>;
  UpdateRoute(request: UpdateRouteRequest): Promise<RoutePlan>;
  UpdatePickupStatus(request: UpdatePickupRequest): Promise<Ack>;
  GetDriver(request: DriverId): Promise<DriverProfile>;
  GetDriverByUserId(request: UserId): Promise<DriverProfile>;
}

export const DriverServiceServiceName = "lastmile.driver.DriverService";
export class DriverServiceClientImpl implements DriverService {
  private readonly rpc: Rpc;
  private readonly service: string;
  constructor(rpc: Rpc, opts?: { service?: string }) {
    this.service = opts?.service || DriverServiceServiceName;
    this.rpc = rpc;
    this.RegisterDriver = this.RegisterDriver.bind(this);
    this.RegisterRoute = this.RegisterRoute.bind(this);
    this.UpdateRoute = this.UpdateRoute.bind(this);
    this.UpdatePickupStatus = this.UpdatePickupStatus.bind(this);
    this.GetDriver = this.GetDriver.bind(this);
    this.GetDriverByUserId = this.GetDriverByUserId.bind(this);
  }
  RegisterDriver(request: RegisterDriverRequest): Promise<DriverProfile> {
    const data = RegisterDriverRequest.encode(request).finish();
    const promise = this.rpc.request(this.service, "RegisterDriver", data);
    return promise.then((data) => DriverProfile.decode(new BinaryReader(data)));
  }

  RegisterRoute(request: RegisterRouteRequest): Promise<RoutePlan> {
    const data = RegisterRouteRequest.encode(request).finish();
    const promise = this.rpc.request(this.service, "RegisterRoute", data);
    return promise.then((data) => RoutePlan.decode(new BinaryReader(data)));
  }

  UpdateRoute(request: UpdateRouteRequest): Promise<RoutePlan> {
    const data = UpdateRouteRequest.encode(request).finish();
    const promise = this.rpc.request(this.service, "UpdateRoute", data);
    return promise.then((data) => RoutePlan.decode(new BinaryReader(data)));
  }

  UpdatePickupStatus(request: UpdatePickupRequest): Promise<Ack> {
    const data = UpdatePickupRequest.encode(request).finish();
    const promise = this.rpc.request(this.service, "UpdatePickupStatus", data);
    return promise.then((data) => Ack.decode(new BinaryReader(data)));
  }

  GetDriver(request: DriverId): Promise<DriverProfile> {
    const data = DriverId.encode(request).finish();
    const promise = this.rpc.request(this.service, "GetDriver", data);
    return promise.then((data) => DriverProfile.decode(new BinaryReader(data)));
  }

  GetDriverByUserId(request: UserId): Promise<DriverProfile> {
    const data = UserId.encode(request).finish();
    const promise = this.rpc.request(this.service, "GetDriverByUserId", data);
    return promise.then((data) => DriverProfile.decode(new BinaryReader(data)));
  }
}

interface Rpc {
  request(service: string, method: string, data: Uint8Array): Promise<Uint8Array>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };